Add files via upload
This is a class file and a running file for charged particles in EM fields. There are options to have the electric and magnetic field to vary over time and the option to have the magnetic field oscillate in time. It produces a 3D plot of the particles path throughout the simulation.