Add files via upload
In my cyclotron file I can build a bunch of lots of particles with slightly different starting velocities.

The file then runs through with relativistic updates working.

I then build a dataframe of derived data such as the average position, velocity and energy of the bunch and the spread of energies of the bunch. 

The derived data analysis file then plots graphs for each of these quantities as well as determining the cyclotron frequency as time develops. 

I now need to build all my test files and doc string and caption my code as well as building a readme.txt